python 3 list/range and int-division type fixes

+++ warpDrive/test_buffers.py
@@ -138,7 +138,7 @@ def test_recycling(g_buf=None):
 def test_recycling_with_overlap(g_buf=None):
     buffer_length = 32
     indices = [set(range(buffer_length))]
-    for bi in range(buffer_length / 2, buffer_length):
+    for bi in range(int(buffer_length / 2), buffer_length):
         indices.append(set(range(bi)))
 
     g_buf = gpu_cpu_comparison(buffer_length, indices, g_buf)
